LayoutTests/platform/mac/accessibility/table-roles-hierarchy-expected.txt

No	Country	Capital
1.	Poland	Warsaw
2.	Russia	Moscow
3.	Ukraine	Kiev
All	3 countries	3 capitals



This shows the hierarchy of table roles.

role: AXRole: AXTable
    role: AXRole: AXRow
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
    role: AXRole: AXRow
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
    role: AXRole: AXRow
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
    role: AXRole: AXRow
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
    role: AXRole: AXRow
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
    role: AXRole: AXColumn
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
    role: AXRole: AXColumn
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
    role: AXRole: AXColumn
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
    role: AXRole: AXGroup
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
        role: AXRole: AXCell
            role: AXRole: AXStaticText
PASS successfullyParsed is true

TEST COMPLETE
